Refactor project structure and migrate host and db servers
