+++ .github/CODEOWNERS
@@ -1,2 +1,2 @@
-# Application Developers of TH-Scribes
-* @TacticalHacker/th-scribes-ad-team
+# Application Developers of TH-Backend
+* @TacticalHacker/tactical-hacker-ad-team
